feat:코드 컨벤션 수정
