+++ src/main/java/com/tag/prietag/dto/user/UserRequest.java
@@ -1,8 +1,7 @@
 package com.tag.prietag.dto.user;
 
 import com.tag.prietag.model.User;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.*;
 
 import javax.validation.constraints.NotEmpty;
 
@@ -38,4 +37,43 @@ public class UserRequest {
         @NotEmpty
         private String password;
     }
+
+    @Getter
+    @Setter
+    @AllArgsConstructor
+    @NoArgsConstructor
+    public class UserLoginDTO {
+
+        @NotEmpty
+        private String username; // 수정: nickname -> username
+        private String password;
+        @NotEmpty
+        private String email;
+        @NotEmpty
+        private User.RoleEnum role;
+    }
+
+    @Getter
+    @Setter
+    @AllArgsConstructor
+    @NoArgsConstructor
+    public class UserSaveInDTO {
+        //    private Long kakaoId;
+        private String username; // kakao_152378934567 방식으로 저장
+        private String password;
+        private String email;
+        private User.RoleEnum role;
+
+        private User toEntity() {
+
+            return User.builder()
+                    .username(username)
+                    .password(password)
+                    .email(email)
+                    .role(User.RoleEnum.USER)
+                    .status(true)
+                    .build();
+        }
+    }
+
 }
